+ check if child not delete then parent can not delete

+++ Modules/ChatBot/Http/Controllers/DialogController.php
@@ -69,6 +69,12 @@ class DialogController extends Controller
             $chatbotDialog = ChatbotDialog::where("id", $id)->first();
 
             if ($chatbotDialog) {
+                // check if it has any parent 
+                $hasChild = ChatbotDialog::where("parent_id",$id)->first();
+                if($hasChild) {
+                    return redirect()->back();
+                }
+
                 WatsonManager::deleteDialog($chatbotDialog->id);
                 ChatbotDialogResponse::where("chatbot_dialog_id", $id)->delete();
                 $chatbotDialog->delete();
@@ -447,6 +453,13 @@ class DialogController extends Controller
     {
         $chatbotDialog = ChatbotDialog::find($id);
         if (!empty($chatbotDialog)) {
+
+            // check if it has any parent 
+            $hasChild = ChatbotDialog::where("parent_id",$id)->first();
+            if($hasChild) {
+                return response()->json(["code" => 500 , "error" => "Parent node can not delete before child : {$hasChild->name}"]);
+            }
+
             // delete old values and send new again start
             $responseCondition = $chatbotDialog->parentResponse()->where("response_type", "response_condition")->get();
             if (!$responseCondition->isEmpty()) {
